Major refactorings and rework on key bindings
