+++ internal/loggo/log_view.go
@@ -25,10 +25,9 @@ package loggo
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/aurc/loggo/internal/reader"
-	"regexp"
-	"strings"
 	"time"
+
+	"github.com/aurc/loggo/internal/reader"
 
 	"github.com/aurc/loggo/internal/color"
 	"github.com/aurc/loggo/internal/config"
@@ -56,6 +55,7 @@ type LogView struct {
 	inSlice            []map[string]interface{}
 	globalCount        int64
 	isFollowing        bool
+	hideFilter         bool
 }
 
 func NewLogReader(app *LoggoApp, reader reader.Reader) *LogView {
@@ -138,29 +138,6 @@ func (l *LogView) processSampleForConfig(sampling []map[string]interface{}) {
 	l.app.config = l.config
 }
 
-func (l *LogView) textViewMenuControl(tv *tview.TextView, onFocus func()) *tview.TextView {
-	tv.SetBlurFunc(func() {
-		tv.Highlight("")
-	}).SetInputCapture(func(event *tcell.EventKey) *tcell.EventKey {
-		switch event.Key() {
-		case tcell.KeyEnter:
-			onFocus()
-			return nil
-		case tcell.KeyESC:
-			l.app.SetFocus(l.table)
-			return nil
-		}
-		return event
-	})
-	tv.SetHighlightedFunc(func(added, removed, remaining []string) {
-		if len(removed) == 0 {
-			onFocus()
-		}
-	})
-	//onFocus()
-	return tv
-}
-
 func (l *LogView) makeUIComponents() {
 	l.templateView = NewTemplateView(l.app, false, func() {
 		// Toggle full screen func
@@ -218,24 +195,7 @@ func (l *LogView) makeUIComponents() {
 		}
 	})
 
-	l.app.SetInputCapture(func(event *tcell.EventKey) *tcell.EventKey {
-		switch event.Key() {
-		case tcell.KeyCtrlF:
-			return nil
-		case tcell.KeyCtrlA:
-			go func() {
-				l.showAbout()
-			}()
-			return nil
-		case tcell.KeyCtrlT:
-			l.makeLayoutsWithTemplateView()
-			return nil
-		case tcell.KeyF2:
-			l.toggledFollowing()
-			return nil
-		}
-		return event
-	})
+	l.keyEvents()
 
 	l.linesView = tview.NewTextView().SetDynamicColors(true).SetTextAlign(tview.AlignRight)
 	l.followingView = tview.NewTextView().
@@ -247,121 +207,39 @@ func (l *LogView) makeUIComponents() {
 	l.followingView.SetBlurFunc(func() {
 		l.followingView.Highlight("")
 	})
-	l.navMenu = tview.NewFlex().SetDirection(tview.FlexRow)
-	l.navMenu.
-		SetBackgroundColor(color.ColorBackgroundField).SetBorderPadding(0, 0, 1, 0)
-
-	l.navMenu.
-		AddItem(tview.NewTextView().
-			SetDynamicColors(true).
-			SetText("[purple]━━━━━ Navigation ━━━━━━").
-			SetTextAlign(tview.AlignCenter), 1, 2, false).
-		AddItem(tview.NewTextView().
-			SetDynamicColors(true).
-			SetText("[yellow::b] ↲      [-::-] View"), 1, 2, false).
-		AddItem(tview.NewTextView().
-			SetDynamicColors(true).
-			SetText("[yellow::b] ↓ ↑ ← →[-::-] Navigate"), 1, 3, false).
-		AddItem(l.textViewMenuControl(tview.NewTextView().
-			SetDynamicColors(true).SetRegions(true).
-			SetText(`[yellow::b] g       [-::u]["1"]Top[""]`), func() {
-			l.isFollowing = false
-			l.table.ScrollToBeginning()
-			if len(l.inSlice) > 1 {
-				go l.table.Select(1, 0)
-			}
-		}), 1, 1, false).
-		AddItem(l.textViewMenuControl(tview.NewTextView().
-			SetDynamicColors(true).SetRegions(true).
-			SetText(`[yellow::b] G       [-::u]["1"]Bottom[""]`), func() {
-			l.isFollowing = false
-			l.table.ScrollToEnd()
-			go l.table.Select(len(l.inSlice), 0)
-		}), 1, 2, false).
-		AddItem(l.textViewMenuControl(tview.NewTextView().
-			SetDynamicColors(true).SetRegions(true).
-			SetText(`[yellow::b] ^f      [-::u]["1"]Pg Up[""]`), func() {
-			l.isFollowing = false
-			l.table.InputHandler()(tcell.NewEventKey(tcell.KeyPgUp, '0', 0), func(p tview.Primitive) {})
-		}), 1, 2, false).
-		AddItem(l.textViewMenuControl(tview.NewTextView().
-			SetDynamicColors(true).SetRegions(true).
-			SetText(`[yellow::b] ^b      [-::u]["1"]Pg Down[""]`), func() {
-			l.isFollowing = false
-			l.table.InputHandler()(tcell.NewEventKey(tcell.KeyPgDn, '0', 0), func(p tview.Primitive) {})
-		}), 1, 2, false).
-		AddItem(tview.NewTextView().
-			SetDynamicColors(true).
-			SetText("[purple]━━━━━━ Selection ━━━━━━").
-			SetTextAlign(tview.AlignCenter), 1, 2, false).
-		AddItem(tview.NewTextView().
-			SetDynamicColors(true).
-			SetText("[yellow::b] ⌥ 🖱    [-::-] Horizontal"), 1, 3, false).
-		AddItem(tview.NewTextView().
-			SetDynamicColors(true).
-			SetText("[yellow::b] ⌥ ⌘ 🖱  [-::-] Vertical"), 1, 3, false).
-		AddItem(tview.NewTextView().
-			SetDynamicColors(true).
-			SetText("[purple]━━━━━━━ Stream ━━━━━━━━").
-			SetTextAlign(tview.AlignCenter), 1, 2, false).
-		AddItem(l.followingView, 1, 2, false).
-		AddItem(l.textViewMenuControl(tview.NewTextView().
-			SetDynamicColors(true).SetRegions(true).
-			SetText(`[yellow::b] ^t      [-::u]["1"]Template[""]`), func() {
-			if l.isTemplateViewShown() {
-				// TODO: Find a reliable way to respond to external closure
-			} else {
-				l.makeLayoutsWithTemplateView()
-			}
-		}), 1, 2, false).
-		AddItem(l.textViewMenuControl(tview.NewTextView().
-			SetDynamicColors(true).SetRegions(true).
-			SetText(`[yellow::b] ^f      [-::u]["1"]Local Filter[""]`), func() {
-			if l.isTemplateViewShown() {
-				// TODO: Find a reliable way to respond to external closure
-			} else {
-				l.makeLayoutsWithTemplateView()
-			}
-		}), 1, 2, false).
-		AddItem(tview.NewTextView().
-			SetDynamicColors(true).
-			SetText("[purple]━━━━━ Application ━━━━━").
-			SetTextAlign(tview.AlignCenter), 1, 2, false).
-		AddItem(l.textViewMenuControl(tview.NewTextView().
-			SetDynamicColors(true).SetRegions(true).
-			SetText(`[yellow::b] ^a      [-::u]["1"]About[""]`), func() {
-			go func() {
-				l.showAbout()
-			}()
-		}), 1, 2, false).
-		AddItem(l.textViewMenuControl(tview.NewTextView().SetRegions(true).
-			SetDynamicColors(true).
-			SetText(`[yellow::b] ^c      [-::u]["1"]Quit[""]`), func() {
-			l.app.Stop()
-		}), 0, 1, false).
-		AddItem(tview.NewBox(), 0, 1, false).
-		AddItem(l.linesView, 1, 1, false)
-	l.mainMenu = tview.NewFlex().SetDirection(tview.FlexColumn)
-	l.mainMenu.
-		SetBackgroundColor(color.ColorBackgroundField).SetTitleAlign(tview.AlignCenter)
-	l.mainMenu.
-		AddItem(l.textViewMenuControl(tview.NewTextView().
-			SetDynamicColors(true).SetRegions(true).
-			SetText(`[yellow::b](^t) [-::u]["1"]Template[""]`), func() {
-			if l.isTemplateViewShown() {
-				// TODO: Find a reliable way to respond to external closure
-			} else {
-				l.makeLayoutsWithTemplateView()
-			}
-		}), 0, 2, false).
-		AddItem(l.followingView, 0, 5, false).
-		AddItem(l.textViewMenuControl(tview.NewTextView().SetRegions(true).
-			SetDynamicColors(true).
-			SetText(`[yellow::b](^C) [-::u]["1"]Quit[""]`), func() {
-			l.app.Stop()
-		}), 0, 1, false).
-		AddItem(l.linesView, 0, 3, false)
+	l.populateMenu()
 	l.updateLineView()
+
+	l.filterView = NewFilterView(l.app, false)
+}
+
+func (l *LogView) toggleFilter() {
+	if l.isJsonViewShown() || l.isTemplateViewShown() {
+		l.hideFilter = false
+	} else {
+		l.hideFilter = !l.hideFilter
+	}
+	l.makeLayouts()
+	if !l.hideFilter {
+		go l.app.SetFocus(l.filterView.expressionField)
+	}
+}
+
+func (l *LogView) makeLayouts() {
+	mainContent := tview.NewFlex().SetDirection(tview.FlexColumn).
+		AddItem(l.table, 0, 2, true).
+		AddItem(l.navMenu, 24, 1, false)
+
+	l.Flex.Clear().SetDirection(tview.FlexRow)
+	if !l.hideFilter {
+		l.Flex.AddItem(l.filterView, 4, 2, false).
+			AddItem(NewHorizontalSeparator(color.FieldStyle, LineHThick, "", 0), 1, 2, false)
+	}
+	l.Flex.AddItem(mainContent, 0, 2, false).
+		//AddItem(l.navMenu, 1, 1, false).
+		//AddItem(l.mainMenu, 1, 1, false).
+		SetBackgroundColor(color.ColorBackgroundField)
+	l.app.SetFocus(l.table)
 }
 
 func (l *LogView) showAbout() {
@@ -371,47 +249,21 @@ func (l *LogView) showAbout() {
 	l.app.DismissModal()
 	l.app.Draw()
 }
-func (l *LogView) makeLayouts() {
-	mainContent := tview.NewFlex().SetDirection(tview.FlexColumn).
-		AddItem(l.table, 0, 2, true).
-		AddItem(l.navMenu, 24, 1, false)
-	l.Flex.Clear().SetDirection(tview.FlexRow).
-		AddItem(mainContent, 0, 2, true).
-		//AddItem(l.navMenu, 1, 1, false).
-		//AddItem(l.mainMenu, 1, 1, false).
-		SetBackgroundColor(color.ColorBackgroundField)
-	l.app.SetFocus(l.table)
-}
 
 func (l *LogView) isTemplateViewShown() bool {
 	return l.Flex.GetItemCount() > 0 && l.Flex.GetItem(0) == l.templateView ||
 		l.Flex.GetItemCount() > 1 && l.Flex.GetItem(1) == l.templateView
 }
 
+func (l *LogView) isJsonViewShown() bool {
+	return l.Flex.GetItemCount() > 0 && l.Flex.GetItem(0) == l.jsonView ||
+		l.Flex.GetItemCount() > 1 && l.Flex.GetItem(1) == l.jsonView
+}
+
 func (l *LogView) toggledFollowing() {
 	l.isFollowing = !l.isFollowing
 	l.updateLineView()
 	go l.app.Draw()
-}
-func (l *LogView) updateLineView() {
-	r, _ := l.table.GetSelection()
-	if r > 0 {
-		l.linesView.SetText(
-			fmt.
-				Sprintf(`[yellow::]Line [green::b]%d[yellow::-] ([green::b]%d[yellow::-] lines)`,
-					r,
-					l.globalCount))
-	} else {
-		l.linesView.SetText(
-			fmt.
-				Sprintf(`[green::b]%d[yellow::-] lines`,
-					l.globalCount))
-	}
-	if l.isFollowing {
-		l.followingView.SetText(`[yellow::b] F2 [green::bi]ON[-::-]   [-::u]["1"]Auto-Scroll[""][::-]`)
-	} else {
-		l.followingView.SetText(`[yellow::b] F2 [red::bi]OFF[-::-]  [-::u]["1"]Auto-Scroll[""][::-]`)
-	}
 }
 
 func (l *LogView) makeLayoutsWithJsonView() {
@@ -439,103 +291,4 @@ func (l *LogView) makeLayoutsWithTemplateView() {
 		AddItem(l.mainMenu, 1, 1, false)
 
 	l.app.SetFocus(l.templateView.table)
-}
-
-type LogData struct {
-	tview.TableContentReadOnly
-	logView *LogView
-}
-
-func (d *LogData) GetCell(row, column int) *tview.TableCell {
-	if row == -1 || len(d.logView.inSlice) < row-1 || column == -1 {
-		return nil
-	}
-	if column == 0 {
-		if row == 0 {
-			tc := tview.NewTableCell("[yellow] ☀[white] / [blue]☂ ").
-				SetAlign(tview.AlignCenter).
-				SetBackgroundColor(tcell.ColorBlack).
-				SetSelectable(false)
-			return tc
-		} else {
-			if _, ok := d.logView.inSlice[row-1][config.ParseErr]; ok {
-				tc := tview.NewTableCell(" ︎  ☂   ").
-					SetTextColor(tcell.ColorBlue).
-					SetAlign(tview.AlignCenter).
-					SetBackgroundColor(tcell.ColorBlack)
-				return tc
-			} else {
-				tc := tview.NewTableCell("   ︎☀  ︎ ").
-					SetTextColor(tcell.ColorYellow).
-					SetAlign(tview.AlignCenter).
-					SetBackgroundColor(tcell.ColorBlack)
-				return tc
-			}
-		}
-	}
-	c := d.logView.config
-	if len(c.Keys) == 0 {
-		return nil
-	}
-	k := c.Keys[column-1]
-	tc := tview.NewTableCell(" " + k.Name + " ")
-	if k.MaxWidth > 0 && k.MaxWidth-len(k.Name) >= len(k.Name) {
-		spaces := strings.Repeat(" ", k.MaxWidth-len(k.Name))
-		tc.SetText(" " + k.Name + spaces)
-	}
-	// Set Headers
-	if row == 0 {
-		tc.SetTextColor(tcell.ColorYellow).
-			SetAlign(tview.AlignCenter).
-			SetBackgroundColor(tcell.ColorBlack).
-			SetSelectable(false)
-		return tc
-	}
-	// Set Body Cells
-	cellValue := k.ExtractValue(d.logView.inSlice[row-1])
-	var bgColor, fgColor tcell.Color
-	if len(k.Color.Foreground) == 0 {
-		fgColor = k.Type.GetColor()
-	} else {
-		fgColor = k.Color.GetForegroundColor()
-	}
-	bgColor = k.Color.GetBackgroundColor()
-	if len(k.ColorWhen) > 0 {
-	OUT:
-		for _, kv := range k.ColorWhen {
-			reg, err := regexp.Compile(kv.MatchValue)
-			if err == nil && reg.FindIndex([]byte(cellValue)) != nil {
-				bgColor = kv.Color.GetBackgroundColor()
-				fgColor = kv.Color.GetForegroundColor()
-				break OUT
-			}
-		}
-	}
-	switch k.Type {
-	case config.TypeNumber, config.TypeBool:
-		tc.SetAlign(tview.AlignRight)
-	}
-	if k.MaxWidth > 0 {
-		tc.MaxWidth = k.MaxWidth
-	}
-
-	if k.Name == config.TextPayload {
-		if _, ok := d.logView.inSlice[row-1][config.ParseErr]; ok {
-			fgColor = tcell.ColorBlue
-		}
-	}
-
-	return tc.
-		SetBackgroundColor(bgColor).
-		SetTextColor(fgColor).
-		SetText(fmt.Sprintf("%s", cellValue))
-}
-
-func (d *LogData) GetRowCount() int {
-	return len(d.logView.inSlice) + 1
-}
-
-func (d *LogData) GetColumnCount() int {
-	c := d.logView.config
-	return len(c.Keys) + 1
 }
